Make metric_name not have an underscore at the end of it
Prior to this commit, the metric names for each metric had an
underscore to make the file name come out right and not have
the activemq metrics have a leading underscore.

After this commit, we move the underscore into the curl command
and remove it from each metric name.  Now the activemq metrics
will have a leading underscore.