Give activemq metrics a metric name
Prior to this commit, the activemq metrics were in a file that
didn't indicate they were activemq metrics because we wanted
backwards compatibility with before the other metrics were
added.

After this commit, we give the activemq metrics a metric name
to facilitate future changes of storing multiple metrics in one
file.