Break the appending portion of the curl command out
This commit breaks appending to a file into a sperate variable
to facilitate adding in some additional logic in future commits.

This commit also removes the echo of the curl command that was
placed in each file because it makes reading the resultant data
more complicated.  After this commit the resulting output will
be standard json.